truck_telemetry_dashboard/requirements.txt
flask==3.1.3
flask-socketio==5.6.1
paho-mqtt==2.1.0